Enable arbitrary ruby version using rbenv (also still able to use system ruby)
If $RBENV_VERSION is defined in bootstrap.sh, rbenv is installed for the given
$DEV_USERNAME (if necessary) and the ruby version given by $RBENV_VERSION is built
and installed into DEV_USERNAME's ~/.rbenv (if necessary).

If $RBENV_VERSION is *not* defined, bootstrap.sh and its puppet
apply's behave exactly as the previous commit and use system ruby
only.

All bundle exec or gem exec commands in the puppet manifests will now
excecute either the appropriate rbenv ruby or system ruby command as
the Exec path is now set globally in test.pp.